# src/main/resources/swagger/department.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:9080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/department_service/save": {
      "post": {
        "tags": [
          "department-controller"
        ],
        "operationId": "saveDepartment",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Department"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/department_service/find/{departmentId}": {
      "get": {
        "tags": [
          "department-controller"
        ],
        "operationId": "findByDepartmentId",
        "parameters": [
          {
            "name": "departmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Department"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Department": {
        "type": "object",
        "properties": {
          "departmentId": {
            "type": "integer",
            "format": "int64"
          },
          "departmentName": {
            "type": "string"
          },
          "departmentAddress": {
            "type": "string"
          },
          "departmentCode": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}